requirements/django_requirements.txt
Django==5.1.5
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
pytest==8.3.4
pytest-django==4.9.0
